requirements.txt
py-cord>=2.4.1
python-dotenv>=1.0.0
web3>=5.0.0
